CA-01
Project Overview: With data utilized in this project we can analyze and compare the Air Quality in India and how it compares over the different states. This analysis allows for raw data to be analyzed and visualized for easy interpretation. This dataset provides state, location, so2, no2, rspm, spm, type, agency, and date information. 

File developed in Python utilizing Google Colab, utilizing packages such as pandas, matplotlib, numpy, and scikit-learn. 

Author: Alena Sanchez